+++ src/components/projects/template_lib/dockerfile_templates/index.vue
@@ -50,7 +50,7 @@
     </div>
 </template>
 <script>
-import mixin from '@utils/service_module_mixin'
+import mixin from '@/mixin/serviceModuleMixin'
 import FileAside from './file_aside.vue'
 import FileEditor from './file_editor.vue'
 import FileTree from './file_tree.vue'
