+++ src/components/projects/template_lib/k8s_templates/index.vue
@@ -53,7 +53,7 @@
     </div>
 </template>
 <script>
-import mixin from '@utils/service_module_mixin'
+import mixin from '@/mixin/serviceModuleMixin'
 import FileAside from './file_aside.vue'
 import FileEditor from './file_editor.vue'
 import FileTree from './file_tree.vue'
